👷 Move to Github Actions
This change is a read-across of: https://github.com/share/sharedb/pull/409

It moves us on to Github Actions, because Travis has been pretty slow
lately. Github Actions is also free for public repositories and is (for
now) much faster than Travis.

While we're here, we also update our tested Node and MongoDB versions
in-line with their respective LTS schedules.