💥 Bump `redis` to v4
Fixes https://github.com/share/sharedb-redis-pubsub/issues/19

This is a **BREAKING** change that upgrades the underlying `redis`
client to v4, which had so many [breaking changes][1], that it's
infeasible to maintain backwards-compatibility with v3. If consumers
need to use v3, they are advised to pin their version of
`sharedb-redis-pubsub` to v4.

This change adapts to the breakages:

 - wrap the returned `Promise`s in callbacks
 - actively connect the `client`, which no longer auto-connects
 - move the `message` event handler into the subscription callback
 - adapt to the new call signature of `client.eval()`

[1]: https://github.com/redis/node-redis/blob/HEAD/docs/v3-to-v4.md